pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jobos"
version = "4.0.0"
description = "JobOS — Neurosymbolic Job-Centric Operating System"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    # Core framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",

    # Databases
    "neo4j>=5.25.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # AI/ML Engines (Phase 1 minimal)
    "networkx>=3.4.0",

    # Utilities
    "structlog>=24.4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# Phase 2 engine upgrades
engines = [
    "pymdp>=0.0.8",
    "dowhy>=0.11.0",
    "pgmpy>=0.0.19",
]
# NLP extraction
nlp = [
    "spacy>=3.8.0",
    "openai>=1.50.0",
]
# Document & URL extraction
extract = [
    "trafilatura>=1.6.0",
    "beautifulsoup4>=4.12.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "chardet>=5.0.0",
    "httpx>=0.27.0",
]
# Development
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "hypothesis>=6.112.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/jobos"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
